+++ rubin-cli/src/main.rs
@@ -3,15 +3,16 @@ use cli::*;
 
 use std::io::{self, Write};
 
+use rubin::net::parser::Operation;
 use rubin::net::{client::RubinClient, server::start};
 
-fn prep_input(cmd: &str) -> String {
-    let cmd_split: Vec<&str> = cmd.trim().split(' ').collect();
-    if cmd_split.len() < 2 {
-        return cmd.to_string();
+fn validate_cmd_length(cmds: &Vec<&str>, size: usize) -> bool {
+    if cmds.len() != size {
+        println!("incorrect argument length for operation.\n");
+        return false;
     }
 
-    cmd_split[0].to_owned() + "::" + &cmd_split[1..].join(" ")
+    true
 }
 
 #[tokio::main]
@@ -24,28 +25,74 @@ async fn main() -> io::Result<()> {
         }
         Commands::Cli(args) => {
             let client = RubinClient::new(&args.address, args.port);
-            let mut cmd = String::new();
 
             loop {
-                print!("> ");
+                // The overheads of creating a new string each loop are insignificant
+                let mut cmd = String::new();
+
+                print!("rubin-cli > ");
                 io::stdout().flush()?;
                 io::stdin().read_line(&mut cmd)?;
 
-                if cmd.trim() == "exit" {
+                let trimmed_cmd = cmd.trim();
+                if trimmed_cmd == "exit" {
                     println!("Quitting.");
                     break;
+                } else if trimmed_cmd.is_empty() {
+                    continue;
                 }
-                let prepped_cmd = prep_input(&cmd);
 
-                let response = match client.request(&prepped_cmd).await {
-                    Ok(response) => response,
+                let mut cmd_split: Vec<&str> = cmd.split(' ').collect();
+                let raw_op = &cmd_split.remove(0).trim();
+                let op = Operation::from_string(raw_op);
+
+                let response = match op {
+                    Operation::StringGet => {
+                        if !validate_cmd_length(&cmd_split, 1) {
+                            continue;
+                        }
+                        let key = &cmd_split[0];
+                        client.get_string(key).await
+                    }
+                    Operation::StringSet => {
+                        if !validate_cmd_length(&cmd_split, 2) {
+                            continue;
+                        }
+                        let key = &cmd_split[0];
+                        let value = &cmd_split[1];
+                        client.insert_string(key, value).await
+                    }
+                    Operation::StringRemove => {
+                        if !validate_cmd_length(&cmd_split, 1) {
+                            continue;
+                        }
+                        let key = &cmd_split[0];
+                        client.remove_string(key).await
+                    }
+                    Operation::StringClear => client.clear_strings().await,
+                    Operation::Dump => {
+                        if !validate_cmd_length(&cmd_split, 1) {
+                            continue;
+                        }
+
+                        let path = &cmd_split[0];
+                        client.dump_store(path).await
+                    }
+                    Operation::Error => {
+                        println!("invalid operation: {}\n", raw_op);
+                        continue;
+                    }
+                    _ => continue,
+                };
+
+                let msg = match response {
+                    Ok(msg) => msg,
                     Err(e) => {
                         println!("Unable to connect to the Rubin server: {}", e);
                         break;
                     }
                 };
-                println!("{}\n", response);
-                cmd.clear();
+                println!("{}\n", msg);
             }
         }
     }